Fixing color picker to work on desktop side
